# frontend/openapi.json
{
  "components": {
    "schemas": {
      "AppError": {
        "oneOf": [
          {
            "example": "Internal Server Error",
            "properties": {
              "InternalServerError": {
                "type": "string"
              }
            },
            "required": [
              "InternalServerError"
            ],
            "type": "object"
          },
          {
            "example": "Bad Request",
            "properties": {
              "BadRequest": {
                "type": "string"
              }
            },
            "required": [
              "BadRequest"
            ],
            "type": "object"
          },
          {
            "example": "Not Found",
            "properties": {
              "NotFound": {
                "type": "string"
              }
            },
            "required": [
              "NotFound"
            ],
            "type": "object"
          },
          {
            "example": "Unauthorized",
            "properties": {
              "Unauthorized": {
                "type": "string"
              }
            },
            "required": [
              "Unauthorized"
            ],
            "type": "object"
          },
          {
            "example": "Forbidden",
            "properties": {
              "Forbidden": {
                "type": "string"
              }
            },
            "required": [
              "Forbidden"
            ],
            "type": "object"
          },
          {
            "example": "Storage Error",
            "properties": {
              "Storage": {
                "type": "string"
              }
            },
            "required": [
              "Storage"
            ],
            "type": "object"
          }
        ]
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "description": "A JWT signed with the secret of the server, sent as `Authorization: Bearer <token>` (`Authorization: Token <token>`, which InfluxDB clients send, works too). Each operation lists the scope it needs: `read`, `write`, `delete` or `admin`. Make one with `sensapp generate-token`, or with `POST /api/v1/admin/tokens` and an `admin` token (the Credentials tab of the web UI does). With authentication disabled (`SENSAPP_AUTH_DISABLED`) no token is needed.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Sensor data ingestion, storage, and retrieval platform",
    "license": {
      "identifier": "Apache-2.0",
      "name": "Apache-2.0"
    },
    "title": "sensapp",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/prometheus_remote_write": {
      "post": {
        "description": "Allows you to write data from Prometheus to SensApp.\n\nIt follows the [Prometheus Remote Write specification](https://prometheus.io/docs/concepts/remote_write_spec/).",
        "operationId": "publish_prometheus",
        "parameters": [
          {
            "description": "Content encoding, must be snappy",
            "in": "header",
            "name": "content-encoding",
            "required": true,
            "schema": {
              "format": "snappy",
              "type": "string"
            }
          },
          {
            "description": "Content type, must be application/x-protobuf",
            "in": "header",
            "name": "content-type",
            "required": true,
            "schema": {
              "format": "application/x-protobuf",
              "type": "string"
            }
          },
          {
            "description": "Prometheus Remote Write version, must be 0.1.0",
            "in": "header",
            "name": "x-prometheus-remote-write-version",
            "required": true,
            "schema": {
              "format": "0.1.0",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/x-protobuf": {}
          },
          "description": "Prometheus Remote Write endpoint. [Reference](https://prometheus.io/docs/concepts/remote_write_spec/)"
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            },
            "description": "Bad Request"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppError"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "description": "The storage backend is unavailable, or SensApp is busy writing and sheds the write. A `Retry-After` header (seconds, randomised) means the request was not processed at all and can be sent again after that delay",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before sending the write again",
                "schema": {
                  "format": "int32",
                  "minimum": 0,
                  "type": "integer"
                }
              }
            }
          },
          "504": {
            "description": "The request took longer than SENSAPP_HTTP_WRITE_TIMEOUT_SECONDS (five minutes by default), usually because the storage backend hangs. The batches of samples already written stay stored, so a retry can store them twice"
          }
        },
        "security": [
          {
            "bearer": [
              "write"
            ]
          }
        ],
        "summary": "Prometheus Remote Write API.",
        "tags": [
          "Prometheus"
        ]
      }
    },
...
  },
...
}